Initial folder structure and FTP files
I added a basic folder structure and R script files to be filled in. I also downloaded both public use microdata and the FTP text files from teh BLS website. Because of their size, I added the PUMD files to the .gitignore file. There is a file readme-pumd.txt that describes the files that are needed.