more steps to automate download
took more steps towards automating download and import. now there is a file in the root dir bls_files.csv which is basically a list of all the files on the BLS website that someone might want for this project, along with metadata (such as url folder, desired object name, and eventually the fn to use to read them in). After sourcing setup.R and import_fns.R, try running get_bls_tables() and see if the download works for you.